+++ Extra/Settings/SettingGui.js
@@ -21,6 +21,7 @@ import SoopyOpenGuiEvent from "../../../guimanager/EventListener/SoopyOpenGuiEve
 import { fetch } from "../../Utils/networkUtils"
 import NumberTextBox from "../../../guimanager/GuiElement/NumberTextBox"
 import TextBox from "../../../guimanager/GuiElement/TextBox"
+import BoxWithGear from "../../../guimanager/GuiElement/BoxWithGear"
 import { MESSAGE_PREFIX } from "../../Utils/Utils"
 import Notification from "../../../guimanager/Notification"
 
@@ -31,9 +32,11 @@ class SettingGui {
      * @param {RenderContext} renderContext 
      * @param {MapRenderer} mapRenderer
      */
-    constructor(defaultSettings, fakeDungeon, renderContext, mapRenderer) {
+    constructor(currentSettings, fakeDungeon, renderContext, mapRenderer) {
         this.gui = new SoopyGui()
-        this.defaultSettings = defaultSettings
+        this.currentSettings = currentSettings
+
+        this.addSidebarElement = this.addSidebarElement.bind(this) //BatChamp I LOVE JS this. CONTEXT!
 
         this.gui.setOpenCommand("bettermap")
         this.gui.setOpenCommand("bm")
@@ -53,26 +56,40 @@ class SettingGui {
         this.changelogData = undefined
         this.mainpage = new SoopyGuiElement().setLocation(0, 0, 1, 1).setScrollable(true)
         this.changelog = new SoopyGuiElement().setLocation(1, 0, 1, 1).setScrollable(true)
+        this.moreSettingsPage = new SoopyGuiElement().setLocation(1, 0, 1, 1)
         this.howToUse = new SoopyGuiElement().setLocation(-1, 0, 1, 1).setScrollable(true)
 
         this.mainSidebar.addChild(this.mainpage)
         this.mainSidebar.addChild(this.changelog)
+        this.mainSidebar.addChild(this.moreSettingsPage)
         this.mainSidebar.addChild(this.howToUse)
+
+        this.onSettingChangeFunctions = []
+
+        this.closeMoreSettingsGui = () => { }
 
         this.y = 0.05
 
-        this.mainpage.addChild(new TextWithArrow().setText("§0Changelog").setLocation(0.675, 0, 0.3, 0.05).addEvent(new SoopyMouseClickEvent().setHandler(() => {
+        this.changelogButton = new TextWithArrow().setText("§0Changelog").setLocation(0.675, 0, 0.3, 0.05).addEvent(new SoopyMouseClickEvent().setHandler(() => {
             this.howToUse.location.location.x.set(-2, 250)
             this.mainpage.location.location.x.set(-1, 250)
             this.changelog.location.location.x.set(0, 250)
 
             this.generateChangelog()
-        })))
-        this.mainpage.addChild(new TextWithArrow().setText("§0How To Use").setLocation(0.025, 0, 0.3, 0.05).setDirectionRight(false).addEvent(new SoopyMouseClickEvent().setHandler(() => {
+        }))
+        this.mainpage.addChild(this.changelogButton)
+
+        this.howToUseButton = new TextWithArrow().setText("§0How To Use").setLocation(0.025, 0, 0.3, 0.05).setDirectionRight(false).addEvent(new SoopyMouseClickEvent().setHandler(() => {
             this.howToUse.location.location.x.set(0, 250)
             this.mainpage.location.location.x.set(1, 250)
             this.changelog.location.location.x.set(2, 250)
-        })))
+        }))
+        this.mainpage.addChild(this.howToUseButton)
+
+        this.settingsPageButton = new TextWithArrow().setText("§0Settings").setLocation(0.025, 0, 0.3, 0.05).setDirectionRight(false).addEvent(new SoopyMouseClickEvent().setHandler(() => {
+            this.closeMoreSettingsGui()
+        }))
+        this.moreSettingsPage.addChild(this.settingsPageButton)
 
         this.howToUse.addChild(new TextWithArrow().setText("§0Settings").setLocation(0.675, 0, 0.3, 0.05).addEvent(new SoopyMouseClickEvent().setHandler(() => {
             this.howToUse.location.location.x.set(-1, 250)
@@ -86,7 +103,7 @@ class SettingGui {
         // TITLE
         this.addSidebarElement(new SoopyTextElement().setText("§0BetterMap Settings").setMaxTextScale(3))
 
-        this.addToggle("Map enabled", "showMap", this.defaultSettings.showMap)
+        this.addToggle("Map enabled", "showMap", this.currentSettings.showMap)
 
         this.addDropdown("Map Style", {
             "legalmap": "Legal Map",
@@ -104,7 +121,7 @@ class SettingGui {
 
 
         // Location edit gui
-        let editLocationGui = new LocationGui(this.defaultSettings.posX ?? 0, this.defaultSettings.posY ?? 0, (this.defaultSettings.size ?? 150) / 100, () => this.gui.open()).onChange(val => {
+        let editLocationGui = new LocationGui(this.currentSettings.posX ?? 0, this.currentSettings.posY ?? 0, (this.currentSettings.size ?? 150) / 100, () => this.gui.open()).onChange(val => {
             this.changed("posX", val.x)
             this.changed("posY", val.y)
             this.changed("size", val.scale * 100)
@@ -117,20 +134,33 @@ class SettingGui {
         this.addSidebarElement() //adds 2 gaps (button from above diddnt get one added automatically + seperating setting areas)
         this.addCategory("Style Settings")
 
+        if (this.currentSettings.tickStyle === "secrets_underhead") {
+            this.currentSettings.tickStyle = "secrets"
+            this.currentSettings.tickStyle_secrets_overHead = false
+            Client.scheduleTask(10, () => {
+                this.changed("tickStyle", "secrets")
+                this.changed("tickStyle_secrets_overHead", false)
+            })
+        }
         this.addDropdown("Tick Style", {
             "default": "NEU Map",
             "hypixel": "Hypixel",
             "tenios": "tenios",
-            "secrets": "Secrets Found (Above head)",
-            "secrets_underhead": "Secrets Found (Under head)",
-        }, "tickStyle", this.defaultSettings.tickStyle)
+            "secrets": "Secrets Found",
+        }, "tickStyle", this.currentSettings.tickStyle)
+
+        this.addGear(() => {
+            return this.currentSettings.tickStyle === "secrets"
+        }, (elm) => {
+            elm.addToggle("Show secrets over player heads", "tickStyle_secrets_overHead", this.currentSettings.tickStyle_secrets_overHead)
+        })
 
 
         this.addDropdown("Puzzle Style", {
             "none": "None",
             "text": "Text",
             "icon": "Icon"
-        }, "puzzleNames", this.defaultSettings.puzzleNames)
+        }, "puzzleNames", this.currentSettings.puzzleNames)
 
         //this.addToggle("Border around heads", "headBorder", this.defaultSettings.headBorder)
         Client.scheduleTask(0.5 * 20, () => {
@@ -146,31 +176,34 @@ class SettingGui {
             "none": "None",
             "single": "Single Color",
             "class-color": "Class Colors"
-        }, "headBorder", this.defaultSettings.headBorder)[1].setLore(["All border colors can be changed in the config file"])
+        }, "headBorder", this.currentSettings.headBorder)[1].setLore(["All border colors can be changed in the config file"])
 
-        /* perhaps some day the stuff will be added to be able to do this
-        if (renderContext.settings.headBorder == "single"){
-            this.addColorSelector("Player Border Color", "singleBorderColor", this.defaultSettings.singleBorderColor)
-        } else if (renderContext.settings.headBorder == "class-color"){
-            this.addColorSelector("Healer Border Color", "healerColor", this.defaultSettings.healerColor)
-            this.addColorSelector("Mage Border Color", "mageColor", this.defaultSettings.mageColor)
-            this.addColorSelector("Berserk Border Color", "bersColor", this.defaultSettings.bersColor)
-            this.addColorSelector("Archer Border Color", "healerColor", this.defaultSettings.healerColor)
-            this.addColorSelector("Tank Border Color", "tankColor", this.defaultSettings.tankColor)
-        }*/
+        this.addGear(() => {
+            return this.currentSettings.headBorder === "class-color" || this.currentSettings.headBorder === "single"
+        }, (elm) => {
+            if (this.currentSettings.headBorder == "single") {
+                elm.addColorSelector("Player Border Color", "singleBorderColor", this.currentSettings.singleBorderColor)
+            } else if (this.currentSettings.headBorder == "class-color") {
+                elm.addColorSelector("Healer Border Color", "healerColor", this.currentSettings.healerColor)
+                elm.addColorSelector("Mage Border Color", "mageColor", this.currentSettings.mageColor)
+                elm.addColorSelector("Berserk Border Color", "bersColor", this.currentSettings.bersColor)
+                elm.addColorSelector("Archer Border Color", "healerColor", this.currentSettings.healerColor)
+                elm.addColorSelector("Tank Border Color", "tankColor", this.currentSettings.tankColor)
+            }
+        })
 
         this.addDropdown("Player names on map", {
             "never": "Never",
             "leap": "Holding Leaps",
             "always": "Always"
-        }, "playerNames", this.defaultSettings.playerNames)
+        }, "playerNames", this.currentSettings.playerNames)
 
-        this.addSlider("Head Scale", "headScale", this.defaultSettings.headScale || 8, 2, 15)
-        this.addSlider("Icon Scale", "iconScale", this.defaultSettings.iconScale || 10, 2, 15)
+        this.addSlider("Head Scale", "headScale", this.currentSettings.headScale || 8, 2, 15)
+        this.addSlider("Icon Scale", "iconScale", this.currentSettings.iconScale || 10, 2, 15)
 
-        this.addColorSelector("Map Border Color", "mapBorderColor", this.defaultSettings.mapBorderColor)
-        this.addColorSelector("Map Color", "mapBackgroundColor", this.defaultSettings.mapBackgroundColor)
-        this.addColorSelector("Extra Info Color", "extraInfoBackroundColor", this.defaultSettings.extraInfoBackroundColor)
+        this.addColorSelector("Map Border Color", "mapBorderColor", this.currentSettings.mapBorderColor)
+        this.addColorSelector("Map Color", "mapBackgroundColor", this.currentSettings.mapBackgroundColor)
+        this.addColorSelector("Extra Info Color", "extraInfoBackroundColor", this.currentSettings.extraInfoBackroundColor)
 
         this.addCategory("Secret info Settings")
 
@@ -178,38 +211,44 @@ class SettingGui {
             "none": "None",
             "legalmap": "LegalMap",
             "simplified": "Simplified"
-        }, "scoreInfoUnderMap", this.defaultSettings.scoreInfoUnderMap)
+        }, "scoreInfoUnderMap", this.currentSettings.scoreInfoUnderMap)
+
+        this.addGear(() => {
+            return this.currentSettings.scoreInfoUnderMap === "simplified"
+        }, (elm) => {
+            elm.addToggle("Show 'Mimic' text before cross/tick", "scoreInfoUnderMap_simplified_showMimicText", this.currentSettings.scoreInfoUnderMap_simplified_showMimicText)[1].setLore(["If this is disabled it will still show wether mimic has been killed", "It just wont show the text before the indicator"])
+        })
 
         this.addCategory("Tab Info")
 
-        this.addToggle("Show current Secret total", "tabSecretCount", this.defaultSettings.tabSecretCount)[1].setLore(["Change the secrets found number in tab to also show total secrets in dungeon"])
+        this.addToggle("Show current Secret total", "tabSecretCount", this.currentSettings.tabSecretCount)[1].setLore(["Change the secrets found number in tab to also show total secrets in dungeon"])
 
-        this.addToggle("Show current Crypt total", "tabCryptCount", this.defaultSettings.tabCryptCount)[1].setLore(["Change the crypts found number in tab to also show total crypts in dungeon"])
+        this.addToggle("Show current Crypt total", "tabCryptCount", this.currentSettings.tabCryptCount)[1].setLore(["Change the crypts found number in tab to also show total crypts in dungeon"])
 
-        this.addToggle("Show Mimic Status", "tabMimic", this.defaultSettings.tabMimic)[1].setLore(["Add a line to tab displaying wether the minic has been killed"])
+        this.addToggle("Show Mimic Status", "tabMimic", this.currentSettings.tabMimic)[1].setLore(["Add a line to tab displaying wether the minic has been killed"])
 
-        this.addToggle("Fix Score in Scoreboard", "fixScore", this.defaultSettings.fixScore)[1].setLore(["Replaces the score in the Sidebar-Scoreboard with the correct score"])
+        this.addToggle("Fix Score in Scoreboard", "fixScore", this.currentSettings.fixScore)[1].setLore(["Replaces the score in the Sidebar-Scoreboard with the correct score"])
 
         this.addCategory("Other Settings")
 
-        this.addToggle("Hide map in Boss", "hideInBoss", this.defaultSettings.hideInBoss)
-        this.addToggle("Show tabs on map", "showTabs", this.defaultSettings.showTabs)[1].setLore(["Shows tabs at the top of the map that can be clicked", "eg to look at the dungeon map when in boss"])
+        this.addToggle("Hide map in Boss", "hideInBoss", this.currentSettings.hideInBoss)
+        this.addToggle("Show tabs on map", "showTabs", this.currentSettings.showTabs)[1].setLore(["Shows tabs at the top of the map that can be clicked", "eg to look at the dungeon map when in boss"])
 
         this.addDropdown("Current room info next to map", {
             "none": "None",
             "left": "Left of map",
             "right": "Right of map"
-        }, "currentRoomInfo", this.defaultSettings.currentRoomInfo)[1].setLore(["Shows the same info that would be shown when hovering over a room"])
+        }, "currentRoomInfo", this.currentSettings.currentRoomInfo)[1].setLore(["Shows the same info that would be shown when hovering over a room"])
 
-        this.addToggle("Box wither doors", "boxDoors", this.defaultSettings.boxDoors)[1].setLore(["Not esp, loads door locations from map."])
+        this.addToggle("Box wither doors", "boxDoors", this.currentSettings.boxDoors)[1].setLore(["Not esp, loads door locations from map."])
 
-        this.addToggle("Force paul +10 score", "forcePaul", this.defaultSettings.forcePaul)[1].setLore(["Paul score bonus will get auto-detected when paul is mayor", "But it wont be auto detected from jerry-paul"])
+        this.addToggle("Force paul +10 score", "forcePaul", this.currentSettings.forcePaul)[1].setLore(["Paul score bonus will get auto-detected when paul is mayor", "But it wont be auto detected from jerry-paul"])
 
-        this.addToggle("Dungeon clear breakdown", "clearedRoomInfo", this.defaultSettings.clearedRoomInfo)[1].setLore(["Shows the cleared room count and specific rooms in chat when the dungeon ends"])
+        this.addToggle("Dungeon clear breakdown", "clearedRoomInfo", this.currentSettings.clearedRoomInfo)[1].setLore(["Shows the cleared room count and specific rooms in chat when the dungeon ends"])
 
-        this.addToggle("Show secret waypoints (scuffed)", "showSecrets", this.defaultSettings.showSecrets)[1].setLore(["Syncs between bettermap users"])
+        this.addToggle("Show secret waypoints (scuffed)", "showSecrets", this.currentSettings.showSecrets)[1].setLore(["Syncs between bettermap users"])
 
-        this.addToggle("Spirit leap overlay", "spiritLeapOverlay", this.defaultSettings.spiritLeapOverlay)[1].setLore(["You can click on player heads in overlay!", "Most people probs wont like the design though."])
+        this.addToggle("Spirit leap overlay", "spiritLeapOverlay", this.currentSettings.spiritLeapOverlay)[1].setLore(["You can click on player heads in overlay!", "Most people probs wont like the design though."])
 
         this.addSidebarElement(new ButtonWithArrow().setText("&0Load api key from other mods").addEvent(new SoopyMouseClickEvent().setHandler(() => {
             findKey(key => {
@@ -218,9 +257,9 @@ class SettingGui {
         })), 0.3, 0.4, 0.075)
         this.addSidebarElement() //adds a gap because the button diddnt auto add one
 
-        this.apiKeySetting = this.addString("Api key", "apiKey", this.defaultSettings.apiKey)[0]
+        this.apiKeySetting = this.addString("Api key", "apiKey", this.currentSettings.apiKey)[0]
 
-        this.addToggle("Show dev info", "devInfo", this.defaultSettings.devInfo)
+        this.addToggle("Show dev info", "devInfo", this.currentSettings.devInfo)
 
         // END OF SETTINGS
 
@@ -281,6 +320,58 @@ class SettingGui {
         this.apiKeySetting.setText(key)
 
         this.changed("apiKey", key)
+    }
+
+    /**
+     * Will add a gear to the right of the last added setting
+     * @param {function():Boolean} shouldShowFun
+     * @param {function(CustomSettingsBuilder)} generateGuiElementFun
+     * @example ```js
+     * this.addGear(()=>{
+     *      //Return true/false wether to show the gear (based of selected settings)
+     *      //eg this will only show the button if you have spirit leap overlay enabled
+     *      return this.currentSettings.spiritLeapOverlay
+     * }, (elm)=>{
+     *      //add settings to the the soopy gui element to show when the user clicks on the gear
+     *      //eg
+     *      elm.addToggle("Spirit leap overlay", "spiritLeapOverlay", this.defaultSettings.spiritLeapOverlay)[1].setLore(["You can click on player heads in overlay!", "Most people probs wont like the design though."])
+     * })
+     * ```
+     */
+    addGear(shouldShowFun = () => { }, generateGuiElementFun = () => { }) {
+        let shouldShowStart = shouldShowFun()
+
+        let button = new BoxWithGear().setLocation(0.905, this.y - 0.05 - 0.075 / 2, 0.075, 0.075)
+        this.mainpage.addChild(button)
+
+        button.visable = shouldShowStart
+
+        button.addEvent(new SoopyMouseClickEvent().setHandler(() => {
+            //Open button
+            let element = new CustomSettingsBuilder(this)
+            generateGuiElementFun(element)
+
+            this.moreSettingsPage.clearChildren()
+
+            this.moreSettingsPage.addChild(element.getElement())
+            this.moreSettingsPage.addChild(this.settingsPageButton)
+
+            this.mainpage.location.location.x.set(-1, 250)
+            this.moreSettingsPage.location.location.x.set(0, 250)
+
+            this.closeMoreSettingsGui = () => {
+                this.mainpage.location.location.x.set(0, 250)
+                this.moreSettingsPage.location.location.x.set(1, 250)
+            }
+        }))
+
+        this.onSettingChangeFunctions.push(() => {
+            let shouldShowNow = shouldShowFun()
+
+            button.visable = shouldShowNow
+        })
+
+        return button
     }
 
     generateChangelog() {
@@ -374,12 +465,12 @@ class SettingGui {
      * @param {String} setting internal name of the setting to control
      * @param {String} defau Default value
      */
-    addDropdown(label, options, setting, defau) {
-        let drop = this.addSidebarElement(new DropDown().setOptions(options).setSelectedOption(this.defaultSettings[setting] ?? defau).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+    addDropdown(label, options, setting, defau, addFun = this.addSidebarElement) {
+        let drop = addFun(new DropDown().setOptions(options).setSelectedOption(this.currentSettings[setting] ?? defau).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
             this.changed(setting, val)
         })), 0.55, 0.35, 0.075)
 
-        return [drop, this.addSidebarElement(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
+        return [drop, addFun(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
     }
 
 
@@ -389,12 +480,12 @@ class SettingGui {
      * @param {String} setting internal name of the setting to control
      * @param {Boolean} defau Default value
      */
-    addToggle(label, setting, defau) {
-        let toggle = this.addSidebarElement(new Toggle().setValue(this.defaultSettings[setting] ?? defau).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+    addToggle(label, setting, defau, addFun = this.addSidebarElement) {
+        let toggle = addFun(new Toggle().setValue(this.currentSettings[setting] ?? defau).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
             this.changed(setting, val)
         })), 0.625, 0.2, 0.05)
 
-        return [toggle, this.addSidebarElement(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
+        return [toggle, addFun(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
     }
 
     /**
@@ -402,12 +493,12 @@ class SettingGui {
      * @param {String} setting internal name of the setting to control
      * @param {String} defau Default value
      */
-    addString(label, setting, defau) {
-        let textBox = this.addSidebarElement(new TextBox().setText(this.defaultSettings[setting] ?? defau).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+    addString(label, setting, defau, addFun = this.addSidebarElement) {
+        let textBox = addFun(new TextBox().setText(this.currentSettings[setting] ?? defau).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
             this.changed(setting, val)
         })), 0.625, 0.2, 0.05)
 
-        return [textBox, this.addSidebarElement(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
+        return [textBox, addFun(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
     }
 
     /**
@@ -416,21 +507,21 @@ class SettingGui {
      * @param {String} setting internal name of the setting to control
      * @param {Object[]} defau Default value
      */
-    addColorSelector(label, setting, defau) {
+    addColorSelector(label, setting, defau, addFun = this.addSidebarElement) {
 
-        let colorPicker = new ColorPicker().setRGBColor(this.defaultSettings[setting][0] ?? defau[0], this.defaultSettings[setting][1] ?? defau[1], this.defaultSettings[setting][2] ?? defau[2]).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+        let colorPicker = new ColorPicker().setRGBColor(this.currentSettings[setting][0] ?? defau[0], this.currentSettings[setting][1] ?? defau[1], this.currentSettings[setting][2] ?? defau[2]).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
             this.changedArr(setting, 0, val[0])
             this.changedArr(setting, 1, val[1])
             this.changedArr(setting, 2, val[2])
         }))
 
         // Slider 3
-        let slider3 = new Slider().setValue(this.defaultSettings[setting][3] ?? defau[3]).setMin(0).setMax(255).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+        let slider3 = new Slider().setValue(this.currentSettings[setting][3] ?? defau[3]).setMin(0).setMax(255).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
             this.changedArr(setting, 3, Math.round(val))
             numberT3.setText(Math.round(val).toString())
         }))
 
-        let numberT3 = new NumberTextBox().setText((this.defaultSettings[setting][3] ?? defau[3]).toString())
+        let numberT3 = new NumberTextBox().setText((this.currentSettings[setting][3] ?? defau[3]).toString())
 
         numberT3.isNumber = isNumber
 
@@ -441,12 +532,12 @@ class SettingGui {
             slider3.setValue(parseInt(val))
         }))
 
-        this.addSidebarElement(colorPicker, 0.5, 0.1, 0.05)
+        addFun(colorPicker, 0.5, 0.1, 0.05)
 
-        this.addSidebarElement(slider3, 0.6, 0.2, 0.05).setLore(["opacity"])
-        this.addSidebarElement(numberT3, 0.8, 0.1, 0.05).setLore(["opacity"])
+        addFun(slider3, 0.6, 0.2, 0.05).setLore(["opacity"])
+        addFun(numberT3, 0.8, 0.1, 0.05).setLore(["opacity"])
 
-        return [[], this.addSidebarElement(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
+        return [[], addFun(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
     }
 
     /**
@@ -457,13 +548,13 @@ class SettingGui {
      * @param {Number} min Minimum value
      * @param {Number} max Maximum value
      */
-    addSlider(label, setting, defau, min, max) {
-        let slider = new Slider().setValue(this.defaultSettings[setting] ?? defau).setMin(min).setMax(max).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+    addSlider(label, setting, defau, min, max, addFun = this.addSidebarElement) {
+        let slider = new Slider().setValue(this.currentSettings[setting] ?? defau).setMin(min).setMax(max).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
             this.changed(setting, Math.round(val))
             numberT.setText(Math.round(val).toString())
         }))
 
-        let numberT = new NumberTextBox().setText((this.defaultSettings[setting] ?? defau).toString())
+        let numberT = new NumberTextBox().setText((this.currentSettings[setting] ?? defau).toString())
 
         numberT.isNumber = (val) => {
             if (val.includes(".")) return false
@@ -478,19 +569,19 @@ class SettingGui {
             slider.setValue(parseInt(val))
         }))
 
-        this.addSidebarElement(slider, 0.55, 0.2, 0.05)
-        this.addSidebarElement(numberT, 0.8, 0.1, 0.05)
-        return [[slider, numberT], this.addSidebarElement(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
+        addFun(slider, 0.55, 0.2, 0.05)
+        addFun(numberT, 0.8, 0.1, 0.05)
+        return [[slider, numberT], addFun(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
     } //Cinda scuffed its a different return type, but trying to keep atleast semi-consistant
 
     /**
      * 
      * @param {String} label The text/name of the category
      */
-    addCategory(label) {
+    addCategory(label, addFun = this.addSidebarElement) {
         let elm = new SoopyTextElement().setText("§7" + label).setMaxTextScale(2)
-        this.addSidebarElement(elm, 0.1, 0.8, 0.06)
-        elm.location.location.y.set(this.y - 0.1 + 0.04)
+        addFun(elm, 0.1, 0.8, 0.06)
+        elm.location.location.y.set(addFun("getYPlease") - 0.1 + 0.04)
         return elm
     }
 
@@ -503,6 +594,9 @@ class SettingGui {
      * @returns {Object} elm
      */
     addSidebarElement(elm = null, x = 0.1, width = 0.8, height = 0.1) {
+        if (elm === "getYPlease") { //OMEGA SCUFFED
+            return this.y
+        }
         if (elm) this.mainpage.addChild(elm.setLocation(x, this.y + 0.05 - height / 2, width, height))
         if (x === 0.1) this.y += 0.1
         return elm
@@ -635,4 +729,91 @@ function isNumber(val) {
     if (val.includes(".")) return false
     val = "" + val; //coerce num to be a string
     return !isNaN(val) && !isNaN(parseInt(val));
+}
+
+class CustomSettingsBuilder {
+    constructor(parent) {
+        this.parent = parent
+
+        this.addSidebarElement = this.addSidebarElement.bind(this) //BatChamp I LOVE JS this. CONTEXT!
+
+        this.guiElement = new SoopyGuiElement().setLocation(0, 0, 1, 1).setScrollable(true)
+
+        this.y = 0.1
+    }
+
+    getElement() {
+        return this.guiElement
+    }
+
+    /**
+     * @param {Object} elm The element being added
+     * @param {Number} x the x pos
+     * @param {Number} width The width of the elm
+     * @param {Number} height The height of the elm
+     * @returns {Object} elm
+     */
+    addSidebarElement(elm = null, x = 0.1, width = 0.8, height = 0.1) {
+        if (elm === "getYPlease") { //OMEGA SCUFFED
+            return this.y
+        }
+        if (elm) this.guiElement.addChild(elm.setLocation(x, this.y + 0.05 - height / 2, width, height))
+        if (x === 0.1) this.y += 0.1
+        return elm
+    }
+
+    /**
+     * @param {String} label The text to go to the left of the dropdown
+     * @param {Object} options {key:value} where key = setting internal value and value = render text
+     * @param {String} setting internal name of the setting to control
+     * @param {String} defau Default value
+     */
+    addDropdown(label, options, setting, defau) {
+        return this.parent.addDropdown(label, options, setting, defau, this.addSidebarElement)
+    }
+
+    /**
+     * @param {String} label The text to go to the left of the dropdown
+     * @param {String} setting internal name of the setting to control
+     * @param {Boolean} defau Default value
+     */
+    addToggle(label, setting, defau) {
+        return this.parent.addToggle(label, setting, defau, this.addSidebarElement)
+    }
+
+    /**
+     * @param {String} label The text to go to the left of the dropdown
+     * @param {String} setting internal name of the setting to control
+     * @param {String} defau Default value
+     */
+    addString(label, setting, defau) {
+        return this.parent.addString(label, setting, defau, this.addSidebarElement)
+    }
+
+    /**
+     * @param {String} label The text to go to the left of the slider
+     * @param {String} setting internal name of the setting to control
+     * @param {Object[]} defau Default value
+     */
+    addColorSelector(label, setting, defau) {
+        return this.parent.addColorSelector(label, setting, defau, this.addSidebarElement)
+    }
+
+    /**
+     * @param {String} label The text to go to the left of the slider
+     * @param {String} setting internal name of the setting to control
+     * @param {Number} defau Default value
+     * @param {Number} min Minimum value
+     * @param {Number} max Maximum value
+     */
+    addSlider(label, setting, defau, min, max) {
+        return this.parent.addSlider(label, setting, defau, min, max, this.addSidebarElement)
+    }
+
+    /**
+     * @param {String} label The text/name of the category
+     */
+    addCategory(label) {
+        return this.parent.addCategory(label, this.addSidebarElement)
+    }
 }