+++ Render/MapRendering/RoomRenderer.js
@@ -196,11 +196,11 @@ class RoomRenderer {
         return;
         if (room.type === Room.SPAWN || room.type === Room.FAIRY) return
 
-        if (context.mapStyle === 'teniosmap' && (room.type !== Room.PUZZLE || ((context.tickStyle === 'secrets' || context.tickStyle === "secrets_underhead") && context.puzzleNames === 'icon'))) {
+        if (context.mapStyle === 'teniosmap' && (room.type !== Room.PUZZLE || (context.tickStyle === 'secrets' && context.puzzleNames === 'icon'))) {
             let text = null;
             if (room.maxSecrets) {
                 text = room.currentSecrets + '/' + room.maxSecrets
-            } else if (context.tickStyle === 'secrets' || context.tickStyle === "secrets_underhead")
+            } else if (context.tickStyle === 'secrets')
                 text = (room.currentSecrets ?? "?") + "/" + (room.maxSecrets ?? "?");
             if (!text) return;
             let color = ''
@@ -227,9 +227,9 @@ class RoomRenderer {
             y = context.posY + y * (context.size - context.borderWidth) + context.borderWidth
 
 
-            if (context.tickStyle === "secrets") Renderer.translate(0, 0, 100)
+            if (context.tickStyle_secrets_overHead) Renderer.translate(0, 0, 100)
             renderLibs.drawStringCenteredShadow(text, x, y, scale)
-        } else if ((context.tickStyle === 'secrets' || context.tickStyle === "secrets_underhead") && (room.type !== Room.PUZZLE || (context.mapStyle !== 'teniosmap' && context.puzzleNames !== 'text'))) {
+        } else if (context.tickStyle === 'secrets' && (room.type !== Room.PUZZLE || (context.mapStyle !== 'teniosmap' && context.puzzleNames !== 'text'))) {
 
             let location = room.components[0]